+++ lib/orb/ast/abstract_node.rb
@@ -3,15 +3,25 @@
 module ORB
   module AST
     class AbstractNode
-      attr_accessor :children, :attributes, :errors
+      attr_accessor :children, :attributes
+      attr_writer :errors
+
+      EMPTY_ARRAY = [].freeze
 
       def initialize(*_args)
         @children = []
-        @errors = []
+      end
+
+      def errors
+        @errors || EMPTY_ARRAY
       end
 
       def add_child(node)
         @children << node
+      end
+
+      def add_error(error)
+        (@errors ||= []) << error
       end
 
       def render(_context)