+++ lib/orb/temple/compiler.rb
@@ -31,33 +31,33 @@ module ORB
       #
       # rubocop:disable Metrics/CyclomaticComplexity
       # rubocop:disable Metrics/PerceivedComplexity
-      def transform(node, context = {})
+      def transform(node)
         if node.is_a?(ORB::AST::RootNode)
-          transform_children(node, context)
+          transform_children(node)
         elsif node.is_a?(ORB::AST::TextNode)
-          transform_text_node(node, context)
+          transform_text_node(node)
         elsif node.is_a?(ORB::AST::PrintingExpressionNode)
-          transform_printing_expression_node(node, context)
+          transform_printing_expression_node(node)
         elsif node.is_a?(ORB::AST::ControlExpressionNode)
-          transform_control_expression_node(node, context)
+          transform_control_expression_node(node)
         elsif node.is_a?(ORB::AST::TagNode) && node.compiler_directives?
-          transform_directives_for_tag_node(node, context)
+          transform_directives_for_tag_node(node)
         elsif node.is_a?(ORB::AST::TagNode) && node.dynamic?
-          transform_dynamic_tag_node(node, context)
+          transform_dynamic_tag_node(node)
         elsif node.is_a?(ORB::AST::TagNode) && node.html_tag?
-          transform_html_tag_node(node, context)
+          transform_html_tag_node(node)
         elsif node.is_a?(ORB::AST::TagNode) && node.component_tag?
-          transform_component_tag_node(node, context)
+          transform_component_tag_node(node)
         elsif node.is_a?(ORB::AST::TagNode) && node.component_slot_tag?
-          transform_component_slot_tag_node(node, context)
+          transform_component_slot_tag_node(node)
         elsif node.is_a?(ORB::AST::BlockNode)
-          transform_block_node(node, context)
+          transform_block_node(node)
         elsif node.is_a?(ORB::AST::PublicCommentNode)
-          transform_public_comment_node(node, context)
+          transform_public_comment_node(node)
         elsif node.is_a?(ORB::AST::PrivateCommentNode)
-          transform_private_comment_node(node, context)
+          transform_private_comment_node(node)
         elsif node.is_a?(ORB::AST::NewlineNode)
-          transform_newline_node(node, context)
+          transform_newline_node(node)
         elsif node.is_a?(ORB::Error)
           runtime_error(node)
         else
@@ -68,128 +68,116 @@ module ORB
       # rubocop:enable Metrics/PerceivedComplexity
 
       # Compile the children of a node and collect the result into a Temple expression
-      def transform_children(node, context)
-        [:multi, *node.children.map { |child| transform(child, context) }]
+      def transform_children(node)
+        [:multi, *node.children.map { |child| transform(child) }]
       end
 
       # Compile a TextNode into a Temple expression
-      def transform_text_node(node, _context)
+      def transform_text_node(node)
         [:static, node.text]
       end
 
       # Compile an PExpressionNode into a Temple expression
-      def transform_printing_expression_node(node, context)
+      def transform_printing_expression_node(node)
         if node.block?
           tmp = @identity.generate(:variable)
           [:multi,
-            # Capture the result of the code in a variable. We can't do
-            # `[:dynamic, code]` because it's probably not a complete
-            # expression (which is a requirement for Temple).
             [
               :block, "#{tmp} = #{node.expression}",
-
-              # Capture the content of a block in a separate buffer. This means
-              # that `yield` will not output the content to the current buffer,
-              # but rather return the output.
-              #
-              # The capturing can be disabled with the option :disable_capture.
-              # Output code in the block writes directly to the output buffer then.
-              # Rails handles this by replacing the output buffer for helpers.
               if @options.fetch(:disable_capture, false)
-                transform_children(node, context)
+                transform_children(node)
               else
-                [:capture, @identity.generate(:variable), transform_children(node, context)]
+                [:capture, @identity.generate(:variable), transform_children(node)]
               end
             ],
-            # Output the content.
             [:escape, true, [:dynamic, tmp]]]
         elsif node.children.any?
-          [:multi, [:escape, true, [:dynamic, node.expression]], transform_children(node, context)]
+          [:multi, [:escape, true, [:dynamic, node.expression]], transform_children(node)]
         else
           [:escape, true, [:dynamic, node.expression]]
         end
       end
 
       # Compile an NPExpressionNode into a Temple expression
-      def transform_control_expression_node(node, context)
+      def transform_control_expression_node(node)
         if node.block?
           tmp = @identity.generate(:variable)
           [:multi,
-            [:block, "#{tmp} = #{node.expression}", transform_children(node, context)]]
+            [:block, "#{tmp} = #{node.expression}", transform_children(node)]]
         elsif node.children.any?
-          [:multi, [:code, node.expression], transform_children(node, context)]
+          [:multi, [:code, node.expression], transform_children(node)]
         else
           [:code, node.expression]
         end
       end
 
       # Compile an HTML TagNode into a Temple expression
-      def transform_html_tag_node(node, context)
-        [:orb, :tag, node.tag, node.attributes, transform_children(node, context)]
+      def transform_html_tag_node(node)
+        [:orb, :tag, node.tag, node.attributes, transform_children(node)]
       end
 
       # Compile a component TagNode into a Temple expression
-      def transform_component_tag_node(node, context)
-        [:orb, :component, node, transform_children(node, context)]
+      def transform_component_tag_node(node)
+        [:orb, :component, node, transform_children(node)]
       end
 
       # Compile a component slot TagNode into a Temple expression
-      def transform_component_slot_tag_node(node, context)
-        [:orb, :slot, node, transform_children(node, context)]
+      def transform_component_slot_tag_node(node)
+        [:orb, :slot, node, transform_children(node)]
       end
 
       # Compile a block node into a Temple expression
-      def transform_block_node(node, context)
+      def transform_block_node(node)
         case node.name
         when :if
-          [:orb, :if, node.expression, transform_children(node, context)]
+          [:orb, :if, node.expression, transform_children(node)]
         when :for
-          [:orb, :for, node.expression, transform_children(node, context)]
+          [:orb, :for, node.expression, transform_children(node)]
         else
           [:static, 'Unknown block node']
         end
       end
 
       # Compile a comment node into a Temple expression
-      def transform_public_comment_node(node, _context)
+      def transform_public_comment_node(node)
         [:html, :comment, [:static, node.text]]
       end
 
       # Compile a private_comment node into a Temple expression
-      def transform_private_comment_node(_node, _context)
+      def transform_private_comment_node(_node)
         [:static, ""]
       end
 
       # Compile a newline node into a Temple expression
-      def transform_newline_node(_node, _context)
+      def transform_newline_node(_node)
         [:newline]
       end
 
       # Compile a tag node with directives
-      def transform_directives_for_tag_node(node, context)
+      def transform_directives_for_tag_node(node)
         # First, process any :if directives
         if_directive = node.directives.fetch(:if, false)
         if if_directive
           node.remove_directive(:if)
           return [:if,
             if_directive,
-            transform(node, context)]
+            transform(node)]
         end
 
         # Second, process any :for directives
         for_directive = node.directives.fetch(:for, false)
         if for_directive
           node.remove_directive(:for)
-          return [:orb, :for, for_directive, transform(node, context)]
+          return [:orb, :for, for_directive, transform(node)]
         end
 
         # Last, render as a dynamic node expression
-        transform(node, context)
+        transform(node)
       end
 
       # Compile a dynamic tag node
-      def transform_dynamic_tag_node(node, context)
-        [:orb, :dynamic, node, transform_children(node, context)]
+      def transform_dynamic_tag_node(node)
+        [:orb, :dynamic, node, transform_children(node)]
       end
 
       # Helper or raising exceptions during compilation